Se ha hecho un cambio de la version de springboot de 4.0.1 a 3.5.8!!!! endpoints de login, registro y check email creados: dependencias añadidas, Excepciones añadidas, DTOs creados y clases de spring security y jwt añadidas correctamente.

+++ Streaming-BackEnd/src/main/java/tfg/streamingbackend/entitys/embeddedids/CancionProductorId.java
@@ -13,8 +13,8 @@ import lombok.*;
 public class CancionProductorId {
 
     @Column(name = "cancion_id")
-    private Integer cancionId;
+    private Long cancionId;
 
     @Column(name = "usuario_id")
-    private Integer usuarioId;
+    private Long usuarioId;
 }
